+++ lib/src/IsolateArguments.dart
@@ -2,10 +2,11 @@ import 'dart:isolate';
 
 /// Isolate scanner arguments
 class IsolateArguments {
-  SendPort sendPort;
-  String host;
-  List<int> ports;
-  Duration timeout;
+  final SendPort sendPort;
+  final String host;
+  final List<int> ports;
+  final Duration timeout;
+  final Duration updateInterval;
 
-  IsolateArguments(this.sendPort, this.host, this.ports, this.timeout);
+  IsolateArguments(this.sendPort, this.host, this.ports, this.timeout, {this.updateInterval = const Duration(seconds: 1)});
 }